+++ server/pbl03api/src/main/java/com/benkyousuru/pbl03api/model/repository/OrderRepository.java
@@ -4,4 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.benkyousuru.pbl03api.model.entity.Order;
 
-public interface OrderRepository extends CrudRepository<Order, Long> {}
+public interface OrderRepository extends CrudRepository<Order, Integer> {
+    
+}
